improved file structure and logging
