+++ package.json
@@ -5,8 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "build": "tsc -p tsconfig.json",
-    "start": "node index.mjs"
+    "start": "node src/index.mjs"
   },
   "keywords": [],
   "author": "",
@@ -16,5 +15,8 @@
     "fastify": "^4.22.1",
     "linkify-it": "^4.0.1",
     "node-fetch": "^3.3.2"
+  },
+  "devDependencies": {
+    "chalk": "^5.3.0"
   }
 }
