+++ src/utils/linkify.mjs
@@ -1,0 +1,5 @@
+import linkifyIt from 'linkify-it';
+
+const linkify = linkifyIt();
+linkify.add('hf:', 'https:');
+export default linkify
